# Conflicts:
#	korender-framework/.gitignore
#	korender-framework/examples/build.gradle.kts
#	korender-framework/examples/src/commonMain/kotlin/AppExample.kt
#	korender-framework/examples/src/commonMain/kotlin/TestExchange.kt
#	korender-framework/examples/src/commonMain/kotlin/infcity/InfiniteCity.kt
#	korender-framework/examples/src/desktopTest/kotlin/com/zakgof/korenderexamples/PerformanceTest.kt
#	korender-framework/examples/src/desktopTest/kotlin/com/zakgof/korenderexamples/TestUtil.kt
#	korender-framework/examples/src/desktopTest/kotlin/com/zakgof/korenderexamples/perf/InstancedRenderables.kt
#	korender-framework/examples/src/desktopTest/kotlin/com/zakgof/korenderexamples/perf/MultipleRenderables.kt
#	korender-framework/gradle.properties
#	korender-framework/gradle/libs.versions.toml

+++ korender-framework/korender/src/commonMain/kotlin/impl/engine/TouchBox.kt
@@ -4,10 +4,10 @@ import com.zakgof.korender.TouchEvent
 import com.zakgof.korender.TouchHandler
 
 internal class TouchBox(
-    private val x: Int,
-    private val y: Int,
-    private val w: Int,
-    private val h: Int,
+    private val x: Float,
+    private val y: Float,
+    private val w: Float,
+    private val h: Float,
     val id: Any?,
     private val handler: TouchHandler
 ) {
